app/requirements.txt
blinker==1.6.2
boto3==1.26.146
botocore==1.29.146
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jwt==1.3.1
MarkupSafe==2.1.3
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.4
